BE-pep/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
motor
PyJWT
tensorflow>=2.11
numpy>=1.21
pandas>=1.3
matplotlib>=3.4
scikit-learn>=0.24
biopython>=1.78
scikit-learn==1.6.1
openpyxl